requirements.txt
requests==2.7.0
beautifulsoup4==4.4.0
redis==2.10.3
python-dateutil==2.4.2
